requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.5.0
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.11.0
atpublic==3.0.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.0
colorama==0.4.5
commonmark==0.9.1
configobj==5.0.6
cryptography==37.0.2
cycler==0.11.0
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.7.0
dpath==2.0.6
dulwich==0.20.43
dvc==2.12.0
dvc-data==0.0.16
dvc-objects==0.0.16
dvc-render==0.0.6
dvclive==0.9.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
grandalf==0.6
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
mailchecker==4.1.17
MarkupSafe==2.1.1
matplotlib==3.5.1
multidict==6.0.2
nanotime==0.5.2
networkx==2.8.4
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
phonenumbers==8.12.50
Pillow==9.1.1
psutil==5.9.1
pyaml==21.10.1
pycparser==2.21
pydot==1.4.2
pygit2==1.9.2
Pygments==2.12.0
pygtrie==2.4.2
pyparsing==3.0.9
python-benedict==0.25.1
python-dateutil==2.8.2
python-fsutil==0.6.1
python-slugify==6.1.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
scmrepo==0.0.25
shortuuid==1.0.9
shtab==1.5.5
six==1.16.0
smmap==5.0.0
tabulate==0.8.10
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
voluptuous==0.13.1
wcwidth==0.2.5
Werkzeug==2.1.2
xmltodict==0.13.0
yarl==1.7.2
zc.lockfile==2.0
nltk==3.6.3
pymongo==3.12.0
Flask==2.1.3
tweepy==3.10.0
streamlit==1.9.0
protobuf~=3.19.0
altair==4.1.0